Update the baseLedgerEndpoint configuration in the Paladin node configurations to include the type and local nodeName.
Update the build.gradle file for the Paladin operator to include a new task for preparing artifacts.

Add a launch.json file for VS Code to include configurations for running the controller, contract generator, template generator, and artifacts generator.

Update the chart versions in the Paladin operator and Paladin operator CRD charts to reflect the changes.

Signed-off-by: dwertent <david.wertenteil@kaleido.io>