Add GitHub Actions workflow to copy images between registries
This commit introduces a new workflow, `copy-images.yaml`, that facilitates the copying of multi-arch images from GitHub Container Registry (GHCR) to DockerHub. The workflow allows for specifying source and target tags, with an option to tag images as "latest". Additionally, it updates the `release-images.yaml` workflow to include a step for invoking the new copy workflow after pushing images to GHCR.

Signed-off-by: David Wertenteil <david.wertenteil@kaleido.io>
Signed-off-by: Ry Jones <ry@linux.com>